PMM-7873 move Backup utils
